collapse .cached files
